Rename OCR environment for Qwen3
