--- src/benchmax/envs/geo3k_vlm/__init__.py
@@ -1,3 +1,0 @@
-from benchmax.envs.geo3k_vlm.geo3k_vlm_env import Geo3KVLMEnv
-
-__all__ = ["Geo3KVLMEnv"]
