+++ src/benchmax/envs/qwen3_ocr/__init__.py
@@ -1,0 +1,3 @@
+from benchmax.envs.qwen3_ocr.qwen3_ocr_env import Qwen3OCREnv
+
+__all__ = ["Qwen3OCREnv"]
